refactor: move interfaces to standalone files (#27)
