--- dna-frontend/src/app/core/constants/canada-provinces.constant.ts
@@ -1,15 +1,0 @@
-export const CANADA_PROVINCES: string[] = [
-  'AB',
-  'BC',
-  'MB',
-  'NB',
-  'NL',
-  'NS',
-  'NT',
-  'NU',
-  'ON',
-  'PE',
-  'QC',
-  'SK',
-  'YT',
-];
