+++ dna-frontend/src/app/core/models/bracket.model.ts
@@ -1,0 +1,6 @@
+export interface Bracket {
+  minIncome: number;
+  taxRate: number;
+  dividendEligible: number;
+  dividendNonEligible: number;
+}
